rename action-for-ai to actions-for-ai: update folder structure, component imports, page metadata, navbar utm, and RelatedEmbeds references; add className prop to SearchInputHomeOptimized for custom styling; update AutomationsClient with margin and spacing adjustments || 1
