+++ src/app/embed/actions-for-ai/AgentDiagram.js
