+++ src/app/embed/actions-for-ai/HeroAurora.js
