+++ src/app/embed/actions-for-ai/HeroContent.js
